Refactor Git commit-msg hook file parser
The function does too much. It gets arguments that are only there to
then construct the Commit struct. Move this Commit object creation to
the function that calls `parse_commit_hook_format` instead.

Add a test for subject trim behavior that was tested in the git module,
but really belongs in the commit module.