Parse commit message trailers in hook method
When Lintje runs with `--hook-message-file` it will now ask Git to parse
the trailer lines and strip them from the to be parsed commit message
body.

It now works the same as in the behavior without this CLI option, that
was added in the previous commit.

This is the last blocker for more accurately supporting trailer lines
and not considering them part of the message body and the rules that
check the message body for things like length and presence of these
trailer lines.