Fix ignored rules detection in commit messages
With the change in 0.10.0 to better support for trailer lines, the
`lintje:disable` lines were silently ignored if they were at the end of
the message body.

Git detects these lines as trailer lines, which is good, but the code
checking for ignored rules wasn't updated.

It will now detect disabled Lintje rules that are present in the commit
message's trailer lines.