Skip MessageSkipBuildTag rule if tag in subject
Skip the MessageSkipBuildTag rule if a SubjectBuildTag issue is already
found.

It's unsure the build tag is a "skip ci" in the subject. Let's first let
the user resolve the SubjectBuildTag issue before hinting at adding a
build tag to the message body. In the next validation
MessageSkipBuildTag can check if a "skip ci" build tag is present or
not.