Use short links for rule help links
Instead of printing to the full docs link, print a link shorthand. I
don't know if the current URL scheme on the docs page will remain. This
URL shorthand will be more flexible to update the links later. It
doesn't have to know the URL structure in the project itself anymore.
There's also less text on the screen so hopefully it feels less busy.

Part of #35